fix(relay): portable int/int64 in Splice loop for arm32 cross-compile
unix.Splice returns `int` on 32-bit Linux (arm, mipsle) and `int64`
elsewhere — Go's stdlib generates the syscall stubs separately per
GOARCH. The Android armeabi-v7a build (arm32) fails to compile with
"mismatched types int64 and int" on the existing splice loop because
`total` is int64 but `written` is int on that target.

Cast both ways at the assignment to keep the function portable:
- remain -= int(written)    // remain is int
- total  += int64(written)  // total is int64

Surfaced now because the gomobile-bind fix in the same PR unblocked
the cross-compile far enough to expose this. The bug existed since
the splice path was added; CI couldn't see it until bind worked.

+++ internal/relay/splice_linux.go
@@ -74,9 +74,12 @@ func spliceOne(srcFD, dstFD int, pipe *splicePair) (int64, error) {
 			remain := n
 			for remain > 0 {
 				written, werr := unix.Splice(pipe.r, nil, dstFD, nil, int(remain), unix.SPLICE_F_MOVE|unix.SPLICE_F_NONBLOCK)
+				// unix.Splice returns int on 32-bit Linux (arm, mipsle) and
+				// int64 elsewhere. Wide-cast both ways for portability — the
+				// gomobile cross-compile to armeabi-v7a otherwise fails.
 				if written > 0 {
-					remain -= written
-					total += written
+					remain -= int(written)
+					total += int64(written)
 				}
 				if werr != nil {
 					return total, werr
